+++ .idea/compiler.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="CompilerConfiguration">
     <annotationProcessing>
+      <profile default="true" name="Default" enabled="true" />
       <profile name="Maven default annotation processors profile" enabled="true">
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
@@ -9,6 +10,7 @@
         <module name="config-server" />
         <module name="account-service" />
         <module name="eureka-server" />
+        <module name="service-common" />
         <module name="ticket-service" />
       </profile>
     </annotationProcessing>
@@ -18,6 +20,7 @@
       <module name="eureka-server" target="11" />
       <module name="microservice" target="11" />
       <module name="parent-service" target="11" />
+      <module name="service-common" target="11" />
       <module name="ticket-service" target="11" />
     </bytecodeTargetLevel>
   </component>
@@ -28,6 +31,7 @@
       <module name="eureka-server" options="-parameters" />
       <module name="microservice" options="-parameters" />
       <module name="parent-service" options="" />
+      <module name="service-common" options="-parameters" />
       <module name="ticket-service" options="-parameters" />
     </option>
   </component>
