+++ .idea/encodings.xml
@@ -4,6 +4,7 @@
     <file url="file://$PROJECT_DIR$/account-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/config-server/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/eureka-server/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/service-common/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/ticket-service/src/main/java" charset="UTF-8" />
   </component>
 </project>
