+++ .idea/libraries/Maven__io_github_openfeign_feign_slf4j_10_4_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.github.openfeign:feign-slf4j:10.4.0">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/github/openfeign/feign-slf4j/10.4.0/feign-slf4j-10.4.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/github/openfeign/feign-slf4j/10.4.0/feign-slf4j-10.4.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/github/openfeign/feign-slf4j/10.4.0/feign-slf4j-10.4.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
