+++ .idea/libraries/Maven__org_hdrhistogram_HdrHistogram_2_1_9.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.hdrhistogram:HdrHistogram:2.1.9">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/hdrhistogram/HdrHistogram/2.1.9/HdrHistogram-2.1.9.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/hdrhistogram/HdrHistogram/2.1.9/HdrHistogram-2.1.9-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/hdrhistogram/HdrHistogram/2.1.9/HdrHistogram-2.1.9-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
