+++ account-service/pom.xml
@@ -55,7 +55,11 @@
             <artifactId>modelmapper-spring</artifactId>
             <version>2.3.0</version>
         </dependency>
-
+        <dependency>
+            <groupId>com.semarslan</groupId>
+            <artifactId>service-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
