+++ account-service/src/main/java/com/semarslan/accountservice/service/AccountService.java
@@ -1,11 +1,8 @@
 package com.semarslan.accountservice.service;
 
-import com.semarslan.accountservice.dto.AccountDto;
-import com.semarslan.accountservice.entity.Account;
+import com.semarslan.client.contract.AccountDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-
-import java.util.List;
 
 public interface AccountService {
