+++ service-common/src/main/java/com/semarslan/client/FeignConfig.java
@@ -1,0 +1,9 @@
+package com.semarslan.client;
+
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@FeignClient
+public class FeignConfig {
+}
