+++ ticket-service/pom.xml
@@ -58,7 +58,11 @@
             <artifactId>spring-boot-starter-data-elasticsearch</artifactId>
             <version>2.4.1</version>
         </dependency>
-
+        <dependency>
+            <groupId>com.semarslan</groupId>
+            <artifactId>service-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
